feat: refactor library page
