+++ .changeset/silent-boats-search.md
@@ -1,0 +1,5 @@
+---
+"impromat": minor
+---
+
+Expand element tags filter bar
