+++ .changeset/warm-bottles-pump.md
@@ -1,0 +1,5 @@
+---
+"impromat": minor
+---
+
+Add bottom navigation bar
