+++ packages/impromat/src/pages/home/HomeContent.tsx
@@ -9,13 +9,13 @@ import {
   IonText,
 } from "@ionic/react";
 import { arrowForwardOutline, information } from "ionicons/icons";
+import { routeLibrary, routeLibraryTab } from "../../routes/library-routes";
 import {
   routeAbout,
   routePrivacyPolicy,
   routeWorkshops,
 } from "../../routes/shared-routes";
-import { Tabs } from "../library/components/LibraryContentComponent";
-import { routeLibrary, routeLibraryTab } from "../../routes/library-routes";
+import { Tabs } from "../library/LibraryPage";
 
 export const HomeContent: React.FC = () => {
   return (
