+++ packages/impromat/src/pages/navigation/responsive-navigation.tsx
@@ -1,1 +1,1 @@
-export const HIDE_MENU_SIZE = "lg";
+export const HIDE_MENU_SIZE = "md";
